backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Binance Integration
python-binance==1.0.19

# Task Queue
celery==5.3.4

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring & Logging
prometheus-fastapi-instrumentator==6.1.0